remove getChain query
