+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/layout.tsx
@@ -15,8 +15,7 @@ import { redirect } from "next/navigation";
 import { defineChain } from "thirdweb";
 import { StarButton } from "../components/client/star-button";
 import { ChainIcon } from "../components/server/chain-icon";
-import { getChain } from "../getChain";
-import { getChainMetadata } from "../utils";
+import { getChain, getChainMetadata } from "../utils";
 import { AddChainToWallet } from "./components/client/add-chain-to-wallet";
 import { ChainPageTabs } from "./components/client/tabs";
 import { ChainCTA } from "./components/server/cta-card";
