+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/pay/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
-import { getChain } from "../../getChain";
+import { getChain } from "../../utils";
 import { InfoCard } from "../components/server/info-card";
 
 export default async function Page(props: { params: { chain_id: string } }) {
