+++ apps/dashboard/src/app/(dashboard)/(chain)/utils.ts
@@ -1,6 +1,8 @@
 import "server-only";
 
 import { THIRDWEB_API_HOST } from "constants/urls";
+// END TEMPORARY
+import { redirect } from "next/navigation";
 import type { ChainCTAProps } from "./[chain_id]/components/server/cta-card";
 import baseBanner from "./temp-assets/base-banner.jpeg";
 import superchainCTABG from "./temp-assets/cta-bg-superchain.png";
@@ -8,7 +10,22 @@ import xaiCTABg from "./temp-assets/cta-bg-xai-connect.png";
 // TEMPORARY
 import xaiBanner from "./temp-assets/xai-banner.jpg";
 import type { ChainMetadataWithServices } from "./types/chain";
-// END TEMPORARY
+
+export async function getChain(
+  chainIdOrSlug: string,
+): Promise<ChainMetadataWithServices> {
+  const res = await fetch(
+    `${THIRDWEB_API_HOST}/v1/chains/${chainIdOrSlug}?includeServices=true`,
+    // revalidate every 15 minutes
+    { next: { revalidate: 15 * 60 } },
+  );
+
+  const result = await res.json();
+  if (!result.data) {
+    redirect("/404");
+  }
+  return result.data as ChainMetadataWithServices;
+}
 
 export async function getChains() {
   const response = await fetch(
